move ready to its own file
